% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmean.r
\name{hmean}
\alias{hmean}
\title{Harmonic Mean}
\usage{
hmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of positive numbers.}

\item{na.rm}{A logical (\code{TRUE} or \code{FALSE}) indicating whether
to strip out missing values before computing.}
}
\value{
A list with values \code{mean} and \code{se}.
}
\description{
Calculates harmonic mean and its standard error.
}
\examples{
  data(BCI_speed_data)
  agoutiData <- subset(BCI_speed_data, species=="agouti")
  hmean(agoutiData$speed)
}
