% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sbm.r
\name{plot.sbm}
\alias{plot.sbm}
\title{Plot size biased distribution}
\usage{
\method{plot}{sbm}(
  mod,
  log = TRUE,
  add = FALSE,
  ppar = list(col = NULL),
  hpar = list(),
  lpar = list(col = "red")
)
}
\arguments{
\item{mod}{A size biased model fit of class \code{\link{sbm}}.}

\item{log}{A logical specifying whether to plot the log transformed or
untransformed data distribution.}

\item{add}{A logical specifying whether to create a new plot, or add a
parametric distribution curve to an existing plot.}

\item{ppar}{A list of plotting parameters passed to plot; accepts
arguments controlling both overall plot appearance and graphical elements
of the histogram.}

\item{hpar}{A list of parameters used when creating the frequency data;
primarily used to determine breaks.}

\item{lpar}{A list of plotting parameters controlling appearance of the
density curve line (if present).}
}
\value{
None.
}
\description{
Plots the frequency distribution of data from a size biased model, with
fitted parametric distribution if the model is parametric and used no
covariates.
}
\examples{
  # Fit and plot log-normal and Weibull models
  data(BCI_speed_data)
  agoutiData <- subset(BCI_speed_data, species=="agouti")
  lmod <- sbm(speed~1, agoutiData, pdf="lnorm")
  wmod <- sbm(speed~1, agoutiData, pdf="weibull")
  plot(lmod, hpar=list(breaks = 40))
  plot(wmod, add=TRUE, lpar=list(col="blue"))
}
