% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbdData}
\alias{sbdData}
\title{A synthetic dataset with fields \code{speed} (a size biased response),
\code{cov1} (a numeric covariate), and \code{cov2} (a factor covariate).}
\format{
A dataframe.
}
\description{
A synthetic dataset with fields \code{speed} (a size biased response),
\code{cov1} (a numeric covariate), and \code{cov2} (a factor covariate).
}
