% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.sbm.r
\name{AIC.sbm}
\alias{AIC.sbm}
\title{Get AIC of size biased model}
\usage{
\method{AIC}{sbm}(mod)
}
\arguments{
\item{mod}{A size biased model fit of class \code{\link{sbm}}.}
}
\value{
A numeric, AIC value, or NULL if the model is non-parametric
 (fitted with \code{pdf = "none"}).
}
\description{
Extracts the Akaike Information Criterion from a fitted parametric size
biased distribution model.
}
\examples{
  data(sbdData)
  lmod1 <- sbm(speed~1, sbdData, pdf="lnorm")
  lmod2 <- sbm(speed~cov1, sbdData, pdf="lnorm")
  AIC(lmod1)
  AIC(lmod2)
}
