% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdData.R
\docType{data}
\name{sbdData}
\alias{sbdData}
\title{A synthetic dataset}
\format{
A dataframe with  100 rows and 3 variables:
\describe{
  \item{speed}{random log normal observations assumed to have been size bias}
  \item{cov1}{a numeric covariate}
  \item{cov2}{a categorical covariate with three levels}
}
}
\description{
A synthetic dataset
}
