% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.r
\name{sbm}
\alias{sbm}
\title{Fit a size biased model}
\usage{
sbm(
  formula,
  data,
  pdf = c("none", "lnorm", "gamma", "weibull"),
  var.range = c(-4, 4),
  trace = FALSE
)
}
\arguments{
\item{formula}{A two-sided formula of the form response ~ covariate + ...}

\item{data}{A dataframe containing the fields named in formula.}

\item{pdf}{A text value naming the probability density function to use.}

\item{var.range}{The range of log variance within which to search when
fitting parametric distributions.}

\item{trace}{Logical defining whether to show fit information when fitting
parametric distributions (passed to \code{\link[bbmle]{mle2}}.}
}
\value{
A list of class \code{\link{sbm}}.
}
\description{
Fits a parametric or non-parametric size biased distribution model to a
positive response variable.
}
\details{
Response values must be strictly positive. To fit a distribution
 without covariates use 1 on the right hand side of the formula. When
 pdf = "none", the harmonic mean and it's standard error are calculated,
 and no covariates can be used.
}
\examples{
  data(sbdData)
  # harmonic mean estimate
  hmod <- sbm(speed~1, sbdData)

  # lognormal estimate with or without a covariate
  lmod1 <- sbm(speed~1, sbdData, pdf="lnorm")
  lmod2 <- sbm(speed~cov1, sbdData, pdf="lnorm")

  # inspect estimates
  hmod$estimate
  lmod1$estimate
}
